Completed all ADR. Completed end-to-end review. Implemented interface critical bug fix.
