+++ CLAUDE.md
@@ -47,7 +47,7 @@ What exists:
 - MCP server on stdio using `io.modelcontextprotocol:kotlin-sdk-server:0.8.4`
 - 8 MCP tools: store_memory, search_memory, claim_memory, session_start, session_end, associate_memories, reclassify_memory, reflect
 - Domain model: Memory, Tier, AssociationType, Association, AttentionScore, SearchFilter
-- Domain messages: Command (9 sealed variants), Event (13 sealed variants), Notification (2 sealed variants)
+- Domain messages: Command (7 sealed variants), Query (2 sealed variants), Event (13 sealed variants), Notification (2 sealed variants)
 - Inbound ports: MemoryPort, LifecyclePort
 - Outbound ports: BackingServicePort, NotificationPort, RelayPort
 - Programmatic logback configuration (stderr for stdio transport)
