+++ site/_adr/README.md
@@ -8,3 +8,4 @@
 | [0004](0004-resilient-storage-array.adoc)       | Resilient Storage Array       | Accepted | 2026-02-28 |
 | [0005](0005-memory-as-aggregate-root.adoc)      | Memory as Aggregate Root      | Accepted | 2026-02-28 |
 | [0006](0006-events-and-commands-as-message-types.adoc) | Events and Commands as Message Types | Accepted | 2026-02-28 |
+| [0007](0007-biological-memory-as-reference-model.adoc) | Biological Memory as Reference Model | Accepted | 2026-02-28 |
